feat: isolate the illustrations and use a registry so that we can style each independently
