+++ packages/twenty-website-new/src/app/(home)/_constants/three-cards-illustration.ts
@@ -22,10 +22,7 @@ export const THREE_CARDS_ILLUSTRATION_DATA: ThreeCardsIllustrationDataType = {
         role: { text: 'Head of Engineering' },
         company: { text: 'Mid-Market Fintech' },
       },
-      illustration: {
-        src: '/illustrations/home/three-cards-illustration/one.glb',
-        title: 'Production grade illustration',
-      },
+      illustration: 'diamond',
     },
     {
       heading: { text: 'AI for rapid iterations', fontFamily: 'sans' },
@@ -37,10 +34,7 @@ export const THREE_CARDS_ILLUSTRATION_DATA: ThreeCardsIllustrationDataType = {
         role: { text: 'Head of Engineering' },
         company: { text: 'Mid-Market Fintech' },
       },
-      illustration: {
-        src: '/illustrations/home/three-cards-illustration/two.glb',
-        title: 'Rapid iteration illustration',
-      },
+      illustration: 'flash',
     },
     {
       heading: { text: 'Control without drag', fontFamily: 'sans' },
@@ -52,10 +46,7 @@ export const THREE_CARDS_ILLUSTRATION_DATA: ThreeCardsIllustrationDataType = {
         role: { text: 'Head of Engineering' },
         company: { text: 'Mid-Market Fintech' },
       },
-      illustration: {
-        src: '/illustrations/home/three-cards-illustration/three.glb',
-        title: 'Control illustration',
-      },
+      illustration: 'lock',
     },
   ],
 };
