+++ packages/twenty-website-new/src/app/partner/_constants/hero.ts
@@ -8,8 +8,6 @@ export const HERO_DATA: HeroIllustrationDataType = {
   body: {
     text: "We're building the #1 open-source CRM, but we can't do it alone. Join our partner ecosystem and grow with us.",
   },
-  illustration: {
-    src: 'https://app.endlesstools.io/embed/1c6c8259-3276-4cf2-84d8-6b7e87e7ec95',
-    title: 'Partner hero illustration',
-  },
+  illustration:
+    'https://app.endlesstools.io/embed/1c6c8259-3276-4cf2-84d8-6b7e87e7ec95',
 };
