+++ packages/twenty-website-new/src/app/why-twenty/_constants/quote.ts
@@ -1,10 +1,7 @@
 import type { QuoteDataType } from '@/sections/Quote/types/QuoteData';
 
 export const QUOTE_DATA: QuoteDataType = {
-  illustration: {
-    src: '/illustrations/why-twenty/quote/quote.glb',
-    title: 'Quote illustration',
-  },
+  illustration: 'quoteQuotes',
   heading: [
     { text: '“In the AI age,\n', fontFamily: 'serif' },
     { text: ' convergence accelerates”', fontFamily: 'sans' },
