--- packages/twenty-website-new/src/design-system/components/Illustration/types/Illustration.ts
@@ -1,4 +1,0 @@
-export type IllustrationType = {
-  src: string;
-  title: string;
-};
