+++ packages/twenty-website-new/src/illustrations/Quote/Quotes.tsx
@@ -1,18 +1,15 @@
 'use client';
 
+import type { IllustrationProps } from '@/illustrations/types';
 import { theme } from '@/theme';
 import { styled } from '@linaria/react';
-import { useEffect, useRef } from 'react';
+import { useLayoutEffect, useRef } from 'react';
 import * as THREE from 'three';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
 
-const FAQ_VISUAL_MODEL_FIT_SCALE = 4;
-const FAQ_VISUAL_ROTATION_RADIANS_PER_SECOND = 0.5;
-const FAQ_VISUAL_WORLD_OFFSET_Y = 0.52;
+const GLB_URL = '/illustrations/why-twenty/quote/quote.glb';
 
-// World spin axis (x, y, z). Camera on +Z: (0,0,1) = spinner in the view plane.
-// Try (0,1,0) or (1,0,0) if motion or framing looks wrong.
-const FAQ_VISUAL_SPIN_AXIS_XYZ: readonly [number, number, number] = [0, 0, 1];
+const QUOTE_VISUAL_MODEL_FIT_SCALE = 3.05;
 
 const scanlineVertexShader = /* glsl */ `
   varying vec3 vWorldPosition;
@@ -103,6 +100,12 @@ function disposeObjectSubtree(root: THREE.Object3D) {
   });
 }
 
+type MeshRestPose = {
+  position: THREE.Vector3;
+  quaternion: THREE.Quaternion;
+  wobblePhase: number;
+};
+
 function applyScanlineMaterials(
   modelRoot: THREE.Object3D,
   lightDirection: THREE.Vector3,
@@ -113,45 +116,45 @@ function applyScanlineMaterials(
     }
 
     sceneObject.material = createScanlineMaterial(lightDirection);
+
+    const mesh = sceneObject;
+    const rest: MeshRestPose = {
+      position: mesh.position.clone(),
+      quaternion: mesh.quaternion.clone(),
+      wobblePhase: mesh.position.y * 4.2 + mesh.position.x * 1.7,
+    };
+    mesh.userData.quoteVisualRest = rest;
   });
 }
 
-const FaqVisualShell = styled.div`
-  bottom: 0;
-  display: block;
-  left: auto;
-  opacity: 0.45;
+const VisualContainer = styled.div`
+  background-color: transparent;
+  border-radius: ${theme.radius(1)};
+  height: min(544px, 70vw);
+  min-height: ${theme.spacing(80)};
   overflow: hidden;
-  pointer-events: none;
-  position: absolute;
-  right: -5%;
-  top: 0;
-  transform: translateY(-11%);
-  width: min(70vw, 750px);
+  position: relative;
+  width: 100%;
 
   @media (min-width: ${theme.breakpoints.md}px) {
-    right: -10%;
-    transform: translateY(-12%);
-    width: min(60vw, 900px);
+    height: 544px;
+    max-width: 646px;
   }
 `;
 
-const FaqVisualCanvasMount = styled.div`
+const CanvasMount = styled.div`
   display: block;
   height: 100%;
+  inset: 0;
   min-width: 0;
+  position: absolute;
   width: 100%;
 `;
 
-export type FaqVisualProps = {
-  src: string;
-  title: string;
-};
-
-export function FaqVisual({ src, title }: FaqVisualProps) {
+export function Quotes(_properties: IllustrationProps) {
   const mountReference = useRef<HTMLDivElement>(null);
 
-  useEffect(() => {
+  useLayoutEffect(() => {
     const container = mountReference.current;
     if (!container) {
       return;
@@ -160,6 +163,8 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
     let cancelled = false;
     let animationFrameId = 0;
 
+    const pointer = { x: 0, y: 0, inside: false };
+    const targetRotation = { x: 0, y: 0 };
     const lightDirectionWorld = new THREE.Vector3(4, 8, 6).normalize();
 
     const scene = new THREE.Scene();
@@ -175,25 +180,21 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
     renderer.setClearColor(0x000000, 0);
     renderer.outputColorSpace = THREE.SRGBColorSpace;
     const canvas = renderer.domElement;
+    canvas.style.cursor = 'pointer';
     canvas.style.display = 'block';
     canvas.style.height = '100%';
+    canvas.style.touchAction = 'none';
     canvas.style.width = '100%';
     container.appendChild(canvas);
 
-    const wheel = new THREE.Group();
-    scene.add(wheel);
-
-    const spinAxisWorld = new THREE.Vector3(
-      FAQ_VISUAL_SPIN_AXIS_XYZ[0],
-      FAQ_VISUAL_SPIN_AXIS_XYZ[1],
-      FAQ_VISUAL_SPIN_AXIS_XYZ[2],
-    ).normalize();
+    const pivot = new THREE.Group();
+    scene.add(pivot);
 
     const clock = new THREE.Clock();
 
     const loader = new GLTFLoader();
     loader.load(
-      src,
+      GLB_URL,
       (gltf) => {
         if (cancelled) {
           disposeObjectSubtree(gltf.scene);
@@ -205,14 +206,13 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
         const center = bounds.getCenter(new THREE.Vector3());
         const size = bounds.getSize(new THREE.Vector3());
         const maxAxis = Math.max(size.x, size.y, size.z, 0.001);
-        const scale = FAQ_VISUAL_MODEL_FIT_SCALE / maxAxis;
+        const scale = QUOTE_VISUAL_MODEL_FIT_SCALE / maxAxis;
 
         modelRoot.position.sub(center);
         modelRoot.scale.setScalar(scale);
 
         applyScanlineMaterials(modelRoot, lightDirectionWorld);
-        wheel.add(modelRoot);
-        wheel.position.y = FAQ_VISUAL_WORLD_OFFSET_Y;
+        pivot.add(modelRoot);
 
         const renderFrame = () => {
           if (cancelled) {
@@ -221,10 +221,59 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
 
           animationFrameId = window.requestAnimationFrame(renderFrame);
           const delta = Math.min(clock.getDelta(), 0.1);
-          wheel.rotateOnWorldAxis(
-            spinAxisWorld,
-            -delta * FAQ_VISUAL_ROTATION_RADIANS_PER_SECOND,
+
+          const rotationDamp = 6.8;
+          const influence = pointer.inside ? 1 : 0.38;
+          targetRotation.y = pointer.x * 0.78 * influence;
+          targetRotation.x = pointer.y * 0.62 * influence;
+
+          pivot.rotation.y = THREE.MathUtils.damp(
+            pivot.rotation.y,
+            targetRotation.y,
+            rotationDamp,
+            delta,
           );
+          pivot.rotation.x = THREE.MathUtils.damp(
+            pivot.rotation.x,
+            targetRotation.x,
+            rotationDamp,
+            delta,
+          );
+
+          const hoverLift = pointer.inside ? 1 : 0;
+          pivot.scale.setScalar(
+            THREE.MathUtils.damp(pivot.scale.x, 1 + hoverLift * 0.12, 7, delta),
+          );
+
+          const mx = pointer.x * (pointer.inside ? 1 : 0.32);
+          const my = pointer.y * (pointer.inside ? 1 : 0.32);
+
+          modelRoot.traverse((sceneObject) => {
+            if (!(sceneObject instanceof THREE.Mesh)) {
+              return;
+            }
+
+            const rest = sceneObject.userData.quoteVisualRest as
+              | MeshRestPose
+              | undefined;
+            if (!rest) {
+              return;
+            }
+
+            const phase = rest.wobblePhase;
+            const wobble = pointer.inside ? 1 : 0.36;
+            sceneObject.position.x =
+              rest.position.x + mx * 0.22 * Math.sin(phase * 1.8);
+            sceneObject.position.z =
+              rest.position.z + my * 0.19 * Math.cos(phase * 1.4);
+            sceneObject.position.y =
+              rest.position.y + (mx + my) * 0.055 * Math.sin(phase * 2.5);
+
+            const twist = (mx * 0.34 + my * 0.24) * wobble * Math.sin(phase);
+            sceneObject.quaternion.copy(rest.quaternion);
+            sceneObject.rotateY(twist);
+          });
+
           renderer.render(scene, camera);
         };
 
@@ -234,6 +283,32 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
       undefined,
     );
 
+    const setPointerFromEvent = (event: PointerEvent) => {
+      const rect = canvas.getBoundingClientRect();
+      const normalizedX = ((event.clientX - rect.left) / rect.width) * 2 - 1;
+      const normalizedY = -(((event.clientY - rect.top) / rect.height) * 2 - 1);
+      pointer.x = THREE.MathUtils.clamp(normalizedX, -1, 1);
+      pointer.y = THREE.MathUtils.clamp(normalizedY, -1, 1);
+    };
+
+    const handlePointerEnter = () => {
+      pointer.inside = true;
+    };
+
+    const handlePointerLeave = () => {
+      pointer.inside = false;
+      pointer.x = 0;
+      pointer.y = 0;
+    };
+
+    const handlePointerMove = (event: PointerEvent) => {
+      setPointerFromEvent(event);
+    };
+
+    canvas.addEventListener('pointerenter', handlePointerEnter);
+    canvas.addEventListener('pointerleave', handlePointerLeave);
+    canvas.addEventListener('pointermove', handlePointerMove);
+
     const handleResize = () => {
       if (!mountReference.current || cancelled) {
         return;
@@ -241,9 +316,6 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
 
       const nextWidth = mountReference.current.clientWidth;
       const nextHeight = mountReference.current.clientHeight;
-      if (nextWidth < 1 || nextHeight < 1) {
-        return;
-      }
       camera.aspect = nextWidth / nextHeight;
       camera.updateProjectionMatrix();
       renderer.setSize(nextWidth, nextHeight);
@@ -254,6 +326,9 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
     return () => {
       cancelled = true;
       window.removeEventListener('resize', handleResize);
+      canvas.removeEventListener('pointerenter', handlePointerEnter);
+      canvas.removeEventListener('pointerleave', handlePointerLeave);
+      canvas.removeEventListener('pointermove', handlePointerMove);
       window.cancelAnimationFrame(animationFrameId);
       disposeObjectSubtree(scene);
       renderer.dispose();
@@ -262,15 +337,11 @@ export function FaqVisual({ src, title }: FaqVisualProps) {
         container.removeChild(canvas);
       }
     };
-  }, [src]);
+  }, []);
 
   return (
-    <FaqVisualShell>
-      <FaqVisualCanvasMount
-        aria-label={title}
-        ref={mountReference}
-        role="img"
-      />
-    </FaqVisualShell>
+    <VisualContainer>
+      <CanvasMount aria-hidden ref={mountReference} />
+    </VisualContainer>
   );
 }
