+++ packages/twenty-website-new/src/illustrations/index.ts
@@ -1,0 +1,8 @@
+export { IllustrationMount } from './illustration-mount';
+export {
+  ILLUSTRATIONS,
+  THREE_CARDS_ILLUSTRATIONS,
+  type IllustrationId,
+  type ThreeCardsIllustrationId,
+} from './illustrations-registry';
+export type { IllustrationProps } from './types';
