+++ packages/twenty-website-new/src/illustrations/types.ts
@@ -1,0 +1,4 @@
+export type IllustrationProps = {
+  stripeColor?: string;
+  'aria-hidden'?: boolean | 'true' | 'false';
+};
