+++ packages/twenty-website-new/src/sections/Faq/types/FaqData.ts
@@ -1,10 +1,10 @@
-import { EyebrowType } from '@/design-system/components/Eyebrow/types/Eyebrow';
-import { HeadingType } from '@/design-system/components/Heading/types/Heading';
-import { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
-import { FaqQuestionType } from '@/sections/Faq/types/FaqQuestion';
+import { type EyebrowType } from '@/design-system/components/Eyebrow/types/Eyebrow';
+import { type HeadingType } from '@/design-system/components/Heading/types/Heading';
+import type { IllustrationId } from '@/illustrations';
+import { type FaqQuestionType } from '@/sections/Faq/types/FaqQuestion';
 
 export type FaqDataType = {
-  illustration: IllustrationType;
+  illustration: IllustrationId;
   eyebrow: EyebrowType;
   heading: HeadingType[];
   questions: FaqQuestionType[];
