+++ packages/twenty-website-new/src/sections/Footer/types/FooterData.ts
@@ -1,11 +1,11 @@
-import type { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
+import type { IllustrationId } from '@/illustrations';
 import type { FooterBottomType } from './FooterBottom';
 import type { FooterNavGroupType } from './FooterNavGroup';
 import type { FooterSocialLinkType } from './FooterSocialLink';
 
 export type FooterDataType = {
   bottom: FooterBottomType;
-  illustration: IllustrationType;
+  illustration: IllustrationId;
   navGroups: FooterNavGroupType[];
   socialLinks: FooterSocialLinkType[];
 };
