+++ packages/twenty-website-new/src/sections/Helped/types/HeadingCard.ts
@@ -1,10 +1,11 @@
 import type { BodyType } from '@/design-system/components/Body/types/Body';
 import type { HeadingType } from '@/design-system/components/Heading/types/Heading';
-import type { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
+import type { IllustrationId } from '@/illustrations';
 
 export type HeadingCardType = {
   icon: string;
-  illustration: IllustrationType & { color: string };
+  illustration: IllustrationId;
+  color: string;
   heading: HeadingType;
   body: BodyType;
 };
