+++ packages/twenty-website-new/src/sections/Hero/components/index.ts
@@ -1,11 +1,11 @@
+import { Product as ProductVisual } from '@/illustrations/Hero/Product';
+import { WhyTwenty as WhyTwentyVisual } from '@/illustrations/Hero/WhyTwenty';
 import { PartnerVisual } from '@/sections/Hero/components/PartnerVisual/PartnerVisual';
-import { ProductVisual } from '@/sections/Hero/components/ProductVisual/ProductVisual';
 import { Body } from './Body/Body';
 import { Cta } from './Cta/Cta';
 import { Heading } from './Heading/Heading';
 import { HomeVisual } from './HomeVisual/HomeVisual';
 import { Root } from './Root/Root';
-import { WhyTwentyVisual } from './WhyTwentyVisual/WhyTwentyVisual';
 
 export const Hero = {
   Root,
