+++ packages/twenty-website-new/src/sections/Hero/types/HeroIllustrationData.ts
@@ -1,6 +1,5 @@
-import { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
-import { HeroBaseDataType } from '@/sections/Hero/types/HeroBaseData';
+import { type HeroBaseDataType } from '@/sections/Hero/types/HeroBaseData';
 
 export type HeroIllustrationDataType = HeroBaseDataType & {
-  illustration: IllustrationType;
+  illustration: string;
 };
