+++ packages/twenty-website-new/src/sections/Plans/types/PlanCard.ts
@@ -1,11 +1,11 @@
-import { HeadingType } from '@/design-system/components/Heading/types/Heading';
-import { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
-import { PlanFeaturesType } from '@/sections/Plans/types/PlanFeatures';
-import { PlanPriceType } from '@/sections/Plans/types/PlanPrice';
+import { type HeadingType } from '@/design-system/components/Heading/types/Heading';
+import type { IllustrationId } from '@/illustrations';
+import { type PlanFeaturesType } from '@/sections/Plans/types/PlanFeatures';
+import { type PlanPriceType } from '@/sections/Plans/types/PlanPrice';
 
 export type PlanCardType = {
   heading: HeadingType;
   price: PlanPriceType;
-  illustration: IllustrationType;
+  illustration: IllustrationId;
   features: PlanFeaturesType;
 };
