+++ packages/twenty-website-new/src/sections/Plans/types/PlansData.ts
@@ -1,6 +1,6 @@
 import type { BodyType } from '@/design-system/components/Body/types/Body';
 import type { HeadingType } from '@/design-system/components/Heading/types/Heading';
-import type { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
+import type { IllustrationId } from '@/illustrations';
 import type { PlanPriceType } from './PlanPrice';
 
 // Same literals as UI toggles when reading cells.
@@ -29,7 +29,7 @@ export type PlansTierType = {
   cells: PlansTierCellsType;
   featuresTitle: BodyType;
   heading: HeadingType;
-  illustration: IllustrationType;
+  illustration: IllustrationId;
 };
 
 export type PlansDataType = {
