+++ packages/twenty-website-new/src/sections/Quote/types/QuoteData.ts
@@ -1,7 +1,7 @@
 import type { HeadingType } from '@/design-system/components/Heading/types/Heading';
-import type { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
+import type { IllustrationId } from '@/illustrations';
 
 export type QuoteDataType = {
-  illustration: IllustrationType;
+  illustration: IllustrationId;
   heading: HeadingType[];
 };
