+++ packages/twenty-website-new/src/sections/Testimonials/components/index.ts
@@ -1,12 +1,12 @@
+import { Hourglass } from '@/illustrations/Testimonials/Hourglass';
+import { Partner as PartnerVisual } from '@/illustrations/Testimonials/Partner';
 import { Carousel } from './Carousel/Carousel';
-import { HomeVisual } from './HomeVisual/HomeVisual';
-import { PartnerVisual } from './PartnerVisual/PartnerVisual';
 import { Root } from './Root/Root';
 import { Separator } from './Separator/Separator';
 
 export const Testimonials = {
   Carousel,
-  HomeVisual,
+  HomeVisual: Hourglass,
   PartnerVisual,
   Root,
   Separator,
