+++ packages/twenty-website-new/src/sections/ThreeCards/types/ThreeCardsIllustrationCard.ts
@@ -1,12 +1,12 @@
-import { BodyType } from '@/design-system/components/Body/types/Body';
-import { HeadingType } from '@/design-system/components/Heading/types/Heading';
-import { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
-import { ThreeCardsIllustrationCardAttributionType } from './ThreeCardsIllustrationCardAttribution';
+import { type BodyType } from '@/design-system/components/Body/types/Body';
+import { type HeadingType } from '@/design-system/components/Heading/types/Heading';
+import type { ThreeCardsIllustrationId } from '@/illustrations';
+import { type ThreeCardsIllustrationCardAttributionType } from './ThreeCardsIllustrationCardAttribution';
 
 export type ThreeCardsIllustrationCardType = {
   heading: HeadingType;
   body: BodyType;
   benefits?: BodyType[];
   attribution?: ThreeCardsIllustrationCardAttributionType;
-  illustration: IllustrationType;
+  illustration: ThreeCardsIllustrationId;
 };
