+++ packages/twenty-website-new/src/sections/WhyTwentyStepper/types/WhyTwentyStepperData.ts
@@ -1,9 +1,9 @@
 import type { BodyType } from '@/design-system/components/Body/types/Body';
 import type { HeadingType } from '@/design-system/components/Heading/types/Heading';
-import type { IllustrationType } from '@/design-system/components/Illustration/types/Illustration';
+import type { IllustrationId } from '@/illustrations';
 
 export type WhyTwentyStepperDataType = {
   heading: HeadingType;
   body: BodyType[];
-  illustration: IllustrationType;
+  illustration: IllustrationId;
 };
